Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> FYI, at jwerner@'s request:
> https://bugs.chromium.org/p/chromium/issues/detail?id=897992#c3
> I went to great pains to avoid an additional dependency in here:
> https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/1298274/
> Fortunately, it was pretty easy to re-implement the basic library support we need. You can probably just move that code out of "arm" and into a common place if you want to use that.

Is depending on libgpiod still a concern? The library is more mature now and seems to be widely deployed.
I could copy your code but it seems to rely on the pin having a name and in our case (WP pin on Kefka) it is unnamed - we only have a known offset from gpiochip base. Of course I can modify it to suit our case but it would be a bit messy to have two copies of almost the same code - it would be cleaner to use the lib.

Patch-set: 2