Update patch set 2
Patch Set 2:

This code is built by upstream coreboot and we keep having random issues there where people are on this or that 5+ year old Ubuntu LTS and have too old versions of this or that dependency. So it's always nice to keep the overall amount of dependencies as low as possible. If there's a serious maintenance burden in continuing to do this manually we can consider it, but as long as it's just about making a small modification to Brian's stuff (e.g. looking up a pin by offset instead of by name or whatever) I would much prefer we stick with that.

Patch-set: 2
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>