vboot: do not request recovery for VB2_REQUEST_* from VB2_TRY()
When the returned value is not an error (such as VB2_REQUEST_*), do not
call vb2api_fail() from VB2_TRY() to request recovery.

During EC sync, instead of explicitly setting VB2_NV_RECOVERY_REQUEST in
nvdata, utilize either vb2api_fail() or VB2_TRY() to request recovery.

In addition, remove the retry of EC RO sync for the following reasons.
EC sync rarely fails, and even if it does, it's not very likely to be a
transient problem that disappears on the next attempt. Besides, the RO
sync is just a debug feature that only people who have a servo attached
and can manually reflash should be using. Therefore, the retry is
removed and hence we no longer need to restore the recovery request in
nvdata.

BRANCH=none
BUG=chromium:1075488
TEST=make runtests

Change-Id: I9ad8e5e0886679a9a342449553170317b010237b
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>