Update patch set 4
Patch Set 4:

That is for sure the right and best approach.

I don't see how we ever to qualify yet another firmware for BSW. (we released two in the total lifetime of the device). The fixes in the kernel/userspace are due to this fact...

Patch-set: 4
Reviewer: Gerrit User 1349498 <1349498@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>