Update patch set 4
Patch Set 4:

Thank you Brian for involvement. We much appreciate it.
As the bug is P0 and need immediate action, we would like to take the no-brainer fix in user-space (thanks for tip BTW).
We can continue discussion with upstream to see how this "gap" problem will finally be handled but this should not stop us to unblock the release and revert it in case some more solid solution will be available.

Patch-set: 4