2crypto: Force correct section flags for .rodata arrays
This is a follow-up fix to CL:2157900. That CL put the problematic
arrays into sections with the right name, but they still didn't set the
right ELF flags for them (specifically the READONLY flag). This made the
assembler unhappy in coreboot, although it still seems to have built
things correctly in the end:

	{standard input}: Assembler messages:
	{standard input}:359: Warning: setting incorrect section
		attributes for .rodata.vb2_hash_names
	{standard input}:369: Warning: setting incorrect section
		attributes for .rodata.vb2_sig_names

This patch throws more ugly hacks at the problem to make it disappear.

BRANCH=None
BUG=None
TEST=Build coreboot for both Arm and x86 boards, confirmed that error
messages are gone and objdump shows intended section flags.

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: Iea93788a13e9fd14d6b5a16626294d5a4b0e5411