Update patch set 2
Patch Set 2: Code-Review+1

(2 comments)

The downside to this is that force could force a dalboz into a trembyle with this, but since force_update is widespread, it has roughly equivalent behavior. 

the other alternatives are quirks or add a really-force-it flag.

Patch-set: 2
Reviewer: Gerrit User 1380689 <1380689@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1