Update patch set 1
Patch Set 1:

Rocco, it looks like you're adding new tests in this CL as well as re-arranging code.  That makes it hard to review the changes in vb2_ui_action_tests.c, since we can't tell what is new, and what is relocated.

Would it be possible to split this into two CLs -- one to re-arrange the code (no code changes), and another to add the extra tests?

Patch-set: 1
CC: Gerrit User 1126709 <1126709@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>