vboot: add screen stack functionality
Add a stack storing previous screen states.  When the user clicks
"Back", revert to the previous state.

In order to deal with the possibility of a UI cycle (repeatedly
selecting the same sequence of screens), which would eventually
use up all available memory, add the check: When switching to a
specific screen which already exists in the stack, pop until that
screen is reached rather than creating a duplicate stack entry.

BUG=b:146399181, b:158256196
TEST=make clean && make runtests
BRANCH=none

Change-Id: I6fbebc2abb11b26d95d4fcf841eb195b3d589396
Signed-off-by: Joel Kitching <kitching@google.com>