vboot/ui: Synchronize mock inputs for vb2_ui_tests
We have two different mock inputs currently, and soon to be three.

Since the input-solving order should not be strictly defined, it is
difficult to write complicated scenario under current approach. We
can not make sure if it shortcuts and ignores certain mock inputs or
not.

This CL synchronize mock inputs with an iteration counter. We can
writing detailed scenario by assigning the inputs iter-by-iter.

BUG=b:146399181, b:156448738
TEST=make clean && make runtests
TEST=make clean && DETACHABLE=1; make runtests
TEST=make clean && PHYSICAL_PRESENCE_KEYBOARD=1; make runtests
BRANCH=none

Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Change-Id: Ia839a6614eb0453b9f64075298a89a2db59d4070