crossystem: x86: switch to VBNV backup using flashrom instead of mosys
Previously, x86 platforms with vboot2 will backup VBNV in SPI flash
using mosys, which will in turn execute flashrom to preform the
underlying operation.

The set of parent CLs to this commit port this functionality from
mosys directly to vboot's host libraries.  This CL switches to use the
new functionality.

(The CL to switch is provided as a separate CL for x86 only so it's an
easy and clean revert should something go wrong.)

BUG=chromium:1032351,chromium:1030473,chromium:789276
BRANCH=none
TEST=On octupus, write VBNV using crossystem and manually inspect
     RW_NVRAM region in SPI flash.

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I9f945dca99ebd394abea1490fa25d3763834bfa1