crossystem: arm: switch to VBNV using flashrom from mosys
Most ARM platforms will store VBNV in SPI flash by calling out to
mosys, which in turn calls out to flashrom.

The set of parent CLs to this commit port this functionality from
mosys directly to vboot's host libraries.  This CL switches to use the
new functionality.

(The CL to switch is provided as a separate CL for ARM only so it's an
easy and clean revert should something go wrong.)

BUG=chromium:1032351,chromium:1030473,chromium:789276
BRANCH=none
TEST=On scarlet, read and write VBNV using crossystem

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I1949522b665170ebeb35f3c46177f1957980d6a3