vboot/ui: pass extra data to vb2ex_display_ui()
Declare struct vb2_ui_data in 2api.h for screen drawing and pass it as
an argument to vb2ex_display_ui().

Currently this struct only contains one field, 'timer_disabled', which
is used in developer mode screen to decide whether to show the related
description on the screen.

BRANCH=none
BUG=b:146399181, b:157871585
TEST=make runtests
TEST=emerge-puff depthcharge

Cq-Depend: chromium:2236550
Change-Id: I48b46838f482bce612106b840476f5f941cc1166
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>