host/lib/flashrom: enable --fast-verify for write operations
We caused a boot-speed regression as we are currently verifying the
entire flash chip after any write.

Flashrom has an option --fast-verify which verifies only the region
written, which is significantly faster.  It also looks like this is
the way mosys used to handle flashrom writes, so we can align with the
old behavior this way.

BUG=chromium:1091903
BRANCH=none
TEST=unit tests, and boot speed regression went away (on octopus)

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: If8d2288cb0c08e8644b6e05f7b174c3c21542f94