firmware/2lib: Some logic to exit after error is displayed
Basically, after displaying an error, we want the user to be able to
exit the error after they press any key.

The basic idea is that if the state.string field is populated, we need
to display the error string.  If it's set to NULL, no string needs to
be displayed.

BUG=b:144969091
BRANCH=None
TEST=boot into dev mode, try to use ctrl-u, observe error on screen
     hit any key and ensure that screen is redrawn w/o error message.
     boot into recovery, try ctrl-d, observe error on screen
     hit any key and ensure that screen is redrawn w/o error meesage.

Signed-off-by: Shelley Chen <shchen@google.com>
Change-Id: I3f398bc91186e259efb3275908520e728c798595