Allow building for non-CrOS environments
There's some code that is architecture specific, but looking at
it, it's code for Chrome OS devices that just happens to be
split along ISA lines.

When compiling with NOCROS=1, these parts are left out, which
allows building on unsupported ISA (such as POWER) and probably
also helps support non-Linux hosts.

Change-Id: I4aaeb56d4521c426a520bc9a1bb49497bec86c35
Signed-off-by: Patrick Georgi <pgeorgi@google.com>