Update patch set 2
Patch Set 2:

I wonder if it would be a better solution to make function stubs that return -1 for everything for unrecognized architectures? The updater can theoretically still be useful on POWER machines (e.g. the parts that allow flashing over a servo shouldn't depend on crossystem results, I think).

Patch-set: 2
CC: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>