firmware/2lib: Add error message for ctrl-U when usb boot is not enabled
Currently, we only have a beep when the user presses ctrl-u.  Adding
an error message to clarify what the error is.  Also adding error code
for legacy boot, which has not landed in tot yet.  Will need to
utilize error in 2ui.c after it lands in tot.  Puff branch will need a
separate change to enable this error on legacy boot as it landed in
puff branch first with a different CL.

BUG=b:164944674
BRANCH=puff, zork
TEST=boot into developer mode
     make sure "boot from external disk" doesn't appear
     press ctrl+U and make sure dialog box with error appears

Cq-Depend: chromium:2359765
Signed-off-by: Shelley Chen <shchen@google.com>
Change-Id: Id8fad79802cb61883758e8c7561163cc17fb0f87