COIL: Add unblocked_terms.txt to block keywords
This CL kicks off COIL for vboot_reference.

Currently the global unblocked_terms.txt are in effect. Since
it's identical to the global blocked_terms.txt, nothing is being
blocked.

This patch adds unblocked_terms.txt, which overrides the global one.
Thus, all changes with a word listed in the global blocked_terms.txt
but not in the local unblocked_terms.txt will be blocked.

BUG=b:165908442, b:163883397
BRANCH=None
TEST=Run repohook.

Cq-Depend: chromium:2369239
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I1618db8e46f2b77397e016c1dfb02f20f66bb3aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/2369360
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Signed-off-by: Martin Roth <martinroth@chromium.org>