2stub: vb2ex_abort()s on host should abort(), not just exit()
Our keyblock and preamble fuzzers have been broken since CL:2353775,
because they don't set up secdata in the context which the tested
functions now depend upon. Unfortunately, we got no alerting about
this... I only just happened to look at ClusterFuzz randomly today and
saw an odd break in execution statistics around that date.

With the new code as it is right now, all fuzzing attempts of the
function end up running vb2ex_abort() from the uninitialized secdata. It
would be great if the fuzzer could report every time it ends up in that
function, but it doesn't seem to do that right now. Supposedly it
reports "crashes", but I guess exit(1) is not a crash, so let's switch
the line to abort() and see if that gets me the reporting behavior I
want (before actually fixing the fuzzer).

BRANCH=None
BUG=None
TEST=None

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: Ie65838f30fa33a7602db253860afc8eeadcac4c6