futility/updater: Add a build time option to preserve ME region
Currently Management Engine(ME) region is preserved only when it is
locked. This leads to some functionality issues during ME update when
the SoC is in S0 state. Audio, shutdown/reboot are some of the
functionalities that are broken after CSME update. Add a build option to
preserve the ME region for all firmware updates.

BUG=b:156991416
BRANCH=None
TEST=Ensure that the ME region is preserved during autoupdate. Ensure
that the suspend/resume sequence works fine after update. Ensure that
the audio functionality is working fine after firmware update. Ensure
that the shutdown/reboot is working fine after firmwware update. Here is
the test sequence:
$(DUT) chromeos-firmwareupdate --mode=autoupdate
$(DUT) suspend_stress_test -c 1
$(DUT) Play some Youtube clip and ensure that the audio works fine
$(DUT) suspend_stress_test -c 1
$(DUT) reboot

Change-Id: I7dd9887f7804374a6e100de16abe5f154e4a213c
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>