futility/updater: Skip preserving CSME region in factory update
On boards where Management Engine (ME) region is marked as preserved,
it is updated post boot after the firmware update. But the scripts to
perform the CSME update post boot are not invoked in factory mode. So
skip preserving the CSME region during the factory update.

BUG=b:156991416
BRANCH=None
TEST=Ensure that the ME region is preserved during autoupdate. Ensure
that the suspend/resume sequence works fine after update. Ensure that
the audio functionality is working fine after firmware update. Ensure
that the shutdown/reboot is working fine after firmwware update. Here is
the test sequence:
$(DUT) chromeos-firmwareupdate --mode=autoupdate
$(DUT) suspend_stress_test -c 1
$(DUT) Play some Youtube clip and ensure that the audio works fine
$(DUT) suspend_stress_test -c 1
$(DUT) reboot

Change-Id: I5cce48d0397f0f692769220d17933dbc24c63104
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>