common.sh: don't let eval terminate cleanup
Eval will terminate the shell on non-zero error code.

"POSIX says that an error in a special built-in utility
(such as eval) should cause the non-interactive shell to terminate"

This is the case and is causing cleanup to terminate android signing
with a non-zero error when it is clear the intent (given the set +e) is
that we should be best effort here.

BUG=chromium:1141907
TEST=unittest and manually on a signer

Change-Id: Ie6374b292c7982371d549b919b44328ea71a09dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/2513228
Reviewed-by: George Engelbrecht <engeg@google.com>
Tested-by: George Engelbrecht <engeg@google.com>
Signed-off-by: Martin Roth <martinroth@chromium.org>