sign_gsc_firmware: refactor and extend to support D2
The major difference between Cr50 and Ti50 signing is that the RW
sections are represented differently: elf files in Cr50 case and ihex
files in Ti50 case.

Other differences include the produced signed final image size and the
offsets of the components in the final image.

The signing script is being updated to figure out all these
differences at run time. A new optional field is introduced in the
signing manifest, the 'generation'. If this field is absent or set to
'h' (for H1), the script proceeds with the Cr50 signing process. If
'generation' is set to 'd' (for D2), the script proceeds with the Ti50
signing process.

Instead of using fixed offsets into the final image, the base
addresses of the components in ihex format are used, the only fixed
value is the base address of the flash image in the chip address space
(0x40000 for H1 vs 0x80000 for D2).

To make this work for H1 the output format of the signed blob produced
by gsc-codesigner is changed from binary to ihex.

BRANCH=none
BUG=b:173049030
TEST=using this script and the signing_istructions.sh module produced
     by the real Cr50 signer was able to produce functional images for
     both Cr50 and Ti50.

(cherry picked from commit fb25edf13402ef3be8935c7d3dc1416efff49c5b)

Change-Id: I845be1101b09c9476fa27fbddb72607dc6cea901
Original-Signed-off-by: Vadim Bendebury <vbendeb@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/2570009
Original-Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Original-Reviewed-by: George Engelbrecht <engeg@google.com>
Original-Auto-Submit: Vadim Bendebury <vbendeb@chromium.org>
Original-Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
GitOrigin-RevId: fb25edf13402ef3be8935c7d3dc1416efff49c5b