crossystem: support driver-level TPM2.0 simulator
After implemented the driver-level TPM2.0 simulator on VM boards, the
mount-encrypted would use the vTPM to encrypted the file system. So we
don't need to remove the mount-encrypted key.
And we would need to revmove the TPM simulator NVChip when we want to
hard reset the TPM on VM.

BUG=b:174807059
BRANCH=none
TEST=crossystem clear_tpm_owner_request=1

Cq-Depend: chromium:2577394
Signed-off-by: Yi Chou <yich@google.com>
Change-Id: Iba2c9b93ed9e558a9163542dfc1fbcb738c1d83d