sign_gsc_firmware: update generated file name
When processing Gsc image singing request the signing server is not
aware of which chip the image is being signed for, the output file
name includes the string @CHIP@ and it is the responsibility of the
actual signing scripts to figure out if the image is for Cr50 or Ti50.

The destination image type is determined based no the signing manifest
contents, this patch add code to replace @CHIP@ with the actual image
type.

BRANCH=none
BUG=b:173049030
TEST=invoked the script to sign a Ti50 image locally, verified that
     the produced signed image file had the expected name.

(cherry picked from commit bc2317695965bb92b8809d9c06327adedcb0653c)

Change-Id: Ib1534ce50e0a44d0ec014e8dbee4e4d85c2082c9
Original-Signed-off-by: Vadim Bendebury <vbendeb@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/2596695
Original-Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Original-Auto-Submit: Vadim Bendebury <vbendeb@chromium.org>
Original-Reviewed-by: George Engelbrecht <engeg@google.com>
Original-Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
GitOrigin-RevId: bc2317695965bb92b8809d9c06327adedcb0653c