futility: update: Add "servo_noreset" opt to tell CCD not to reset
As can be seen in <https://crrev.com/c/2325190>, flashrom can take a
"custom_rst=true" parameter.  Flashrom passes this on to Cr50 and
tells it not to reset.  Among other things, this can be useful to work
around hardware bugs where asserting reset puts us in a bad state.

Let's add a parameter to futility can specify this mode.  We'll enable
this just for "--servo" mode of flashrom and not "--ccd" mode.
Presumably we want "--ccd" to remain simple / autoconfiguring and this
is a bit more of an advanced tweak.

BRANCH=None
BUG=b:177664356
TEST=Can flash coachz with futility

Change-Id: Iab188a92c4eae6373d38fe68ee4107f3e8aa5851
Signed-off-by: Douglas Anderson <dianders@chromium.org>