sign_gsc_firmware: support rewriting filename
When the gsc signer reads the contents of the payload it finds
out if it is a cr50 or ti50 chip. We write the chip type to a
.rename file next to the bin (which has a @CHIP@ in the path) so
that the signer can rename the artifact for placement.

BRANCH=None

BUG=b:179964270
TEST=local signer

(cherry picked from commit ce7171190f1975d1ea3e65a1d46d640ebc243d9e)

Original-Signed-off-by: George Engelbrecht <engeg@google.com>
Change-Id: I0600cb60bb614111802119293ba0c63f2b61c231
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/2728736
Original-Reviewed-by: Mike Frysinger <vapier@chromium.org>
Original-Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Original-Commit-Queue: George Engelbrecht <engeg@google.com>
Original-Tested-by: George Engelbrecht <engeg@google.com>
GitOrigin-RevId: ce7171190f1975d1ea3e65a1d46d640ebc243d9e