vboot/ui: Catch unexpected error from ui actions
Add `default_error_code` to `vb2_screen_info`. If it is set, If it is
set, all the error returned by ui actions will be caught and the
error_code of ui is set to the default_error_code.
Also, if the error is returned by init, reinit, or screen_action,
vb2_ui_screen_back will be called.

BRANCH=none
BUG=b:157625765
TEST=make clean && CC=x86_64-pc-linux-gnu-clang make runtests

Signed-off-by: Chung-Sheng Wu <chungsheng@google.com>
Change-Id: Ib20cd4e99a602bdd567fee89219142de3b4bd342