cgpt: Move all GPT on SPI-NOR infra behind a flag
This piece of code caused serious issues in b/184559695, and it
seems like we have no active users at the moment.

We can punt the decision to remove the code entirely, but for now,
let's stop building and executing it, leaving it to potential
users to fix it up, and refactor/cleanup/test the code.

BRANCH=none
BUG=b:184812319
TEST=`make` does not build `cgpt_wrapper` or any SPI-NOR code.
TEST=`make GPT_SPI_NOR=1` does build it.
TEST=`emerge-$BOARD -v vboot_reference && \
      cros deploy $IP vboot_reference`
     `cgpt find -t kernel` does not print any RW_GPT-related errors
     anymore.

Change-Id: Ie081f372964807caa1b121059288ae761f2f8e43
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>