cros_ec: Use boot mode to check if EC can be trusted
Currently, EC_IN_RW signal is used for the AP to decide whether EC is
running an RO copy or not.

coreboot will set VB2_CONTEXT_TRUSTED based on the EC's boot mode
if supported. It means the EC is in RO and PD is disabled. This patch
makes vb2_allow_recovery check VB2_CONTEXT_TRUSTED prior to starting
recovery mode.

BUG=b:180927027, b:187871195
BRANCH=none
TEST=build

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: Ie7fb337085f9c5f73ee82049d490a719f76a26bb