gbb: Add flag that will enable a firmware image to log to serial
With this flag, coreboot and depthcharge can come as a single image
that selectively logs to the serial console. In production (and most
circumstances really) this will probably be disabled because the serial
console is _sloooow_.

However there are circumstances in development and testing where having
a live insight into what the system is doing is invaluable. Getting that
without reflashing the entire image.serial.bin but just a bit in the GBB
is convenient (during development) to important (during FAFT).

BUG=b:151780766
BRANCH=none
TEST=Not tested yet as this needs to be wired through coreboot and
depthcharge.

Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Change-Id: Ief5568d7ae2d50e8a9f825f230dde2cea8e5d706